refactor: extract jest-utils
