+++ packages/client/test/end-to-end/StreamrClient.test.ts
@@ -1,12 +1,9 @@
 import fs from 'fs'
 import path from 'path'
-
 import { MessageLayer, StreamPartID, toStreamID, toStreamPartID } from 'streamr-client-protocol'
 import { fastPrivateKey, wait } from 'streamr-test-utils'
-
 import {
     getCreateClient,
-    describeRepeats,
     createPartitionedTestStream,
     createStreamPartIterator,
     toStreamDefinition
@@ -16,7 +13,7 @@ import {
     getPublishTestMessages,
     publishManyGenerator
 } from '../test-utils/publish'
-
+import { describeRepeats } from '../test-utils/jest-utils'
 import { StreamrClient } from '../../src/StreamrClient'
 import { Defer } from '../../src/utils/Defer'
 import * as G from '../../src/utils/GeneratorUtils'
