+++ packages/client/test/flakey/Resends.test.ts
@@ -1,5 +1,6 @@
 import { fetchPrivateKeyWithGas, wait } from 'streamr-test-utils'
-import { describeRepeats, createTestStream } from '../test-utils/utils'
+import { createTestStream } from '../test-utils/utils'
+import { describeRepeats } from '../test-utils/jest-utils'
 import { getPublishTestMessages } from '../test-utils/publish'
 import { StreamrClient } from '../../src/StreamrClient'
 import { Defer } from '../../src/utils/Defer'
