+++ packages/client/test/test-utils/jest-utils.ts
@@ -1,0 +1,35 @@
+import { AggregatedError } from '../../src/utils/AggregatedError'
+
+const TEST_REPEATS = (process.env.TEST_REPEATS) ? parseInt(process.env.TEST_REPEATS, 10) : 1
+
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+export function describeRepeats(msg: string, fn: any, describeFn = describe): void {
+    for (let k = 0; k < TEST_REPEATS; k++) {
+        // eslint-disable-next-line no-loop-func
+        describe(msg, () => {
+            describeFn(`test repeat ${k + 1} of ${TEST_REPEATS}`, fn)
+        })
+    }
+}
+
+describeRepeats.skip = (msg: any, fn: any) => {
+    describe.skip(`${msg} – test repeat ALL of ${TEST_REPEATS}`, fn)
+}
+
+describeRepeats.only = (msg: any, fn: any) => {
+    describeRepeats(msg, fn, describe.only)
+}
+
+export function addAfterFn(): (fn: any) => void {
+    const afterFns: any[] = []
+    afterEach(async () => {
+        const fns = afterFns.slice()
+        afterFns.length = 0
+        // @ts-expect-error invalid parameter
+        AggregatedError.throwAllSettled(await Promise.allSettled(fns.map((fn) => fn())))
+    })
+
+    return (fn: any) => {
+        afterFns.push(fn)
+    }
+}
