+++ packages/client/test/unit/GroupKeyPersistence.test.ts
@@ -4,7 +4,8 @@ import { join } from 'path'
 import { tmpdir } from 'os'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { GroupKeyPersistence } from '../../src/encryption/GroupKeyStore'
-import { uid, addAfterFn, describeRepeats, mockContext } from '../test-utils/utils'
+import { uid, mockContext } from '../test-utils/utils'
+import { addAfterFn, describeRepeats } from '../test-utils/jest-utils'
 import LeakDetector from 'jest-leak-detector'
 import { StreamID, toStreamID } from 'streamr-client-protocol'
