+++ packages/client/test/unit/GroupKeyStore.test.ts
@@ -2,7 +2,8 @@ import LeakDetector from 'jest-leak-detector'
 import crypto from 'crypto'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { GroupKeyStore } from '../../src/encryption/GroupKeyStore'
-import { uid, describeRepeats, mockContext } from '../test-utils/utils'
+import { uid, mockContext } from '../test-utils/utils'
+import { describeRepeats } from '../test-utils/jest-utils'
 import { StreamID, toStreamID } from 'streamr-client-protocol'
 
 describeRepeats('GroupKeyStore', () => {
